requirements.txt
pyopenssl
json2html
requests
mysql-connector-python
Flask
gunicorn
pyyaml
flask_basicauth
ssl